use the env var for both
