+++ pkg/util/liveconfig/manager.go
@@ -6,10 +6,7 @@ package liveconfig
 import (
 	"context"
 
-	mgmtcontainerservice "github.com/Azure/azure-sdk-for-go/services/containerservice/mgmt/2021-10-01/containerservice"
 	"k8s.io/client-go/rest"
-
-	"github.com/Azure/ARO-RP/pkg/util/azureclient/mgmt/containerservice"
 )
 
 type Manager interface {
@@ -23,16 +20,8 @@ func NewDev() Manager {
 }
 
 type prod struct {
-	location        string
-	managedClusters containerservice.ManagedClustersClient
-
-	cachedCredentials map[int][]mgmtcontainerservice.CredentialResult
 }
 
-func NewProd(location string, managedClusters containerservice.ManagedClustersClient) Manager {
-	return &prod{
-		location:          location,
-		managedClusters:   managedClusters,
-		cachedCredentials: make(map[int][]mgmtcontainerservice.CredentialResult),
-	}
+func NewProd() Manager {
+	return &prod{}
 }
